requirements.txt
# HTTP / API
requests==2.31.0
pybit==5.6.2

# Telegram
python-telegram-bot==20.7

# Technical Analysis
pandas==2.1.4
ta==0.11.0
numpy==1.26.3

# Scheduling
APScheduler==3.10.4

# Config
python-dotenv==1.0.0

# Utils
colorama==0.4.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
